+++ browser/js/twitch/content/web/consts.mjs
@@ -44,8 +44,7 @@ export const DEFAULT_SETTINGS = {
             quality_down: 'Minus'
         }
     },
-    CHATLOG_FROM_VOD_SETTINGS: {
-        messages_amount: 12
+    CHATLOG_LOADER_SETTINGS: {
     }
 }
 
@@ -61,8 +60,8 @@ export const DEFAULT_STORAGE = {
                         auto_claim_bonus: false
                     },
                     auto_leave_raids: false,
-                    chatlog_from_vod: true,
-                    chatlog_from_vod_settings: DEFAULT_SETTINGS.CHATLOG_FROM_VOD_SETTINGS,
+                    chatlog_loader: true,
+                    chatlog_loader_settings: DEFAULT_SETTINGS.CHATLOG_LOADER_SETTINGS,
                     always_reply: true
                 },
                 player: {