+++ browser/js/twitch/content/web_idle.js
@@ -40,15 +40,15 @@ async function injectScripts(){
     if(auto_leave_raids){
         injectModuleScript(browser.extension.getURL(WEB_DIR + '/raids_auto_leave.js'));
     }
-    let chatlog_from_vod = stor_sync.settings.twitch.chat.chatlog_from_vod;
-    if(chatlog_from_vod){
-        const CLASS_CHATLOG_FROM_VOD = CONSTS.EXTENSION_CLASS_PREFIX + '-chatlog-from-vod';
-        const CLASS_CHATLOG_FROM_VOD_SCRIPT = CLASS_CHATLOG_FROM_VOD + '-script';
+    let chatlog_loader = stor_sync.settings.twitch.chat.chatlog_loader;
+    if(chatlog_loader){
+        const CLASS_CHATLOG_LOADER = CONSTS.EXTENSION_CLASS_PREFIX + '-chatlog-loader';
+        const CLASS_CHATLOG_LOADER_SCRIPT = CLASS_CHATLOG_LOADER + '-script';
         let script = document.createElement('script');
         script.setAttribute('type', 'module');
-        script.id = CLASS_CHATLOG_FROM_VOD_SCRIPT;
-        script.setAttribute('settings', JSON.stringify(stor_sync.settings.twitch.chat.chatlog_from_vod_settings));
-        script.setAttribute('src', browser.extension.getURL(WEB_DIR + '/chatlog_from_vod.js'));
+        script.id = CLASS_CHATLOG_LOADER_SCRIPT;
+        script.setAttribute('settings', JSON.stringify(stor_sync.settings.twitch.chat.chatlog_loader_settings));
+        script.setAttribute('src', browser.extension.getURL(WEB_DIR + '/chatlog_loader.js'));
         document.head.appendChild(script);
     }
     let always_reply = stor_sync.settings.twitch.chat.always_reply;